Add GetCoursesByCategory endpoint in CourseController
Introduced a new asynchronous HTTP GET endpoint `GetCoursesByCategory` in the `CourseController` class. This method accepts an optional `Category` parameter and retrieves courses from the `_courseRepository` based on the provided category. The courses are mapped to `CourseViewModel` objects using `_mapper` and returned as a JSON response with a success flag and the list of course view models.